split backend by paltform, reimplement windows backend
