+++ src/messages.rs
@@ -1,0 +1,28 @@
+use anyhow::Result;
+use bytesize::ByteSize;
+use slint::SharedString;
+
+pub enum BackendCommand {
+    Refresh,
+    Write {
+        space_to_leave: ByteSize,
+        selected_index: usize,
+        keep_local: bool,
+    },
+    Exit,
+}
+
+pub enum BackendWrite {
+    InProgress(u64, u64, &'static str),
+    Completed(Result<()>),
+}
+
+pub enum BackendEvent {
+    RefreshFinished(anyhow::Result<Vec<slint::SharedString>>),
+    Write(BackendWrite),
+}
+
+#[derive(Clone)]
+pub struct SelectOption {
+    pub label: SharedString,
+}
